[refact] axios 호출 함수 api내부 정의
